Разбиение кода js на модули
